fix comment

+++ src/main/java/com/example/demo/config/service/CarService.java
@@ -9,6 +9,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import static java.util.Objects.isNull;
+import static java.util.Objects.nonNull;
+
 @Service
 public class CarService {
 
@@ -23,22 +26,25 @@ public class CarService {
 
     public ResponseEntity<Car> getCarById(Long id) throws Exception {
         Car car = carRepository.findCarById(id);
-        if (car == null){
+        if (isNull(id)){
             LOGGER.info("Car with " + id + " not found");
             throw new CarNotFoundException(String.format("Car with %d not found", id)); // якщо не знайшов кидай помилку
         }
         return new ResponseEntity<>(car, HttpStatus.OK );
     }
 
-    public ResponseEntity<String> deletedCarById(Long id){
-        carRepository.deleteById(id);
-        LOGGER.info("Deleted car");
-        return new ResponseEntity<>("Deleted car", HttpStatus.OK);
+    public ResponseEntity<String> deletedCarById(Long id) throws CarNotFoundException {
+        if (isNull(id)) {
+            carRepository.deleteById(id);
+            LOGGER.info("Deleted car");
+            return new ResponseEntity<>("Deleted car", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("Car with %d not found", id));
     }
 
     public void updateCar(Car newCar) throws CarNotFoundException {
         Car car = carRepository.findCarById(newCar.getId());
-        if (car != null) {
+        if (nonNull(car)) {
             update(car, newCar);
             LOGGER.info("Updated with: " + newCar.getId());
             carRepository.save(car);
@@ -54,8 +60,11 @@ public class CarService {
         car.setEngineCapacity(newCar.getEngineCapacity());
     }
 
-    public ResponseEntity<String> addCar(Car car){
-        carRepository.save(car);
-        return new ResponseEntity<>("Added", HttpStatus.OK);
+    public ResponseEntity<String> addCar(Car car) throws CarNotFoundException {
+        if (nonNull(car)) {
+            carRepository.save(car);
+            return new ResponseEntity<>("Added", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("Car with %d not found", car.getId()));
     }
 }
