+++ src/main/java/com/example/demo/config/service/OrderService.java
@@ -1,6 +1,7 @@
 package com.example.demo.config.service;
 
 import com.example.demo.config.exception.CarNotFoundException;
+import com.example.demo.config.exception.OrderNotFoundException;
 import com.example.demo.config.models.Order;
 import com.example.demo.config.repositories.OrderRepository;
 import org.slf4j.Logger;
@@ -8,6 +9,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
+
+import static java.util.Objects.isNull;
+import static java.util.Objects.nonNull;
 
 @Service
 public class OrderService {
@@ -23,24 +29,32 @@ public class OrderService {
 
     public ResponseEntity<Order> getOrderById(Long id) throws Exception {
         Order order = orderRepository.findOrderById(id);
-        if (order == null){
+        if (isNull(id)){
             LOGGER.info("Order with " + id + " not found");
             throw new CarNotFoundException(String.format("Order with %d not found", id)); // якщо не знайшов кидай помилку
         }
         return new ResponseEntity<>(order, HttpStatus.OK );
     }
 
-    public ResponseEntity<String> deletedOrderById(Long id){
-        orderRepository.deleteById(id);
-        LOGGER.info("Deleted order");
-        return new ResponseEntity<>("Deleted order", HttpStatus.OK);
+    public ResponseEntity<String> deletedOrderById(Long id) throws CarNotFoundException {
+        if (nonNull(id)) {
+            orderRepository.deleteById(id);
+            LOGGER.info("Deleted order");
+            return new ResponseEntity<>("Deleted order", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("Order with %d not found", id));
     }
 
-    public void updateOrder(Order newOrder) {
+    public void updateOrder(Order newOrder) throws OrderNotFoundException {
         Order order = orderRepository.findOrderById(newOrder.getId());
-        update(order, newOrder);
-        LOGGER.info("Updated with: " + newOrder.getId());
-        orderRepository.save(order);
+        if (nonNull(order)) {
+            update(order, newOrder);
+            LOGGER.info("Updated with: " + newOrder.getId());
+            orderRepository.save(order);
+        } else {
+            LOGGER.info("Order with " + newOrder.getId() + "not found");
+            throw new OrderNotFoundException(String.format("Car with %d not found", newOrder.getId()));
+        }
 
     }
 
@@ -49,9 +63,12 @@ public class OrderService {
         order.setCar(newOrder.getCar());
     }
 
-    public ResponseEntity<String> addOrder(Order order){
-        orderRepository.save(order);
-        return new ResponseEntity<>("Added", HttpStatus.OK);
+    public ResponseEntity<String> addOrder(Order order) throws CarNotFoundException {
+        if (nonNull(order)) {
+            orderRepository.save(order);
+            return new ResponseEntity<>("Added", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("Order with %d not found", order.getId()));
     }
 }
