+++ src/main/java/com/example/demo/config/service/UserService.java
@@ -10,6 +10,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import static java.util.Objects.isNull;
+import static java.util.Objects.nonNull;
+
 @Service
 public class UserService {
 
@@ -25,24 +28,32 @@ public class UserService {
 
     public ResponseEntity<User> getUserById(Long id) throws Exception {
         User user = userRepository.findUserById(id);
-        if (user == null){
+        if (isNull(id)){
             LOGGER.info("User with " + id + " not found");
             throw new CarNotFoundException(String.format("User with %d not found", id)); // якщо не знайшов кидай помилку
         }
         return new ResponseEntity<>(user, HttpStatus.OK );
     }
 
-    public ResponseEntity<String> deletedUserById(Long id){
-        userRepository.deleteById(id);
-        LOGGER.info("Deleted user");
-        return new ResponseEntity<>("Deleted user", HttpStatus.OK);
+    public ResponseEntity<String> deletedUserById(Long id) throws CarNotFoundException {
+        if (nonNull(id)) {
+            userRepository.deleteById(id);
+            LOGGER.info("Deleted user");
+            return new ResponseEntity<>("Deleted user", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("User with %d not found", id));
     }
 
-    public void updateUser(User newUser) {
+    public void updateUser(User newUser) throws CarNotFoundException {
         User user = userRepository.findUserById(newUser.getId());
-        update(user, newUser);
-        LOGGER.info("Updated with: " + newUser.getId());
-        userRepository.save(user);
+        if (nonNull(newUser)) {
+            update(user, newUser);
+            LOGGER.info("Updated with: " + newUser.getId());
+            userRepository.save(user);
+        }else {
+            LOGGER.info("Order with " + newUser.getId() + " not found");
+            throw new CarNotFoundException(String.format("User with %d not found", newUser.getId()));
+        }
 
     }
 
@@ -52,8 +63,12 @@ public class UserService {
        user.setLastName(newUser.getLastName());
     }
 
-    public ResponseEntity<String> addUser(User user){
-        userRepository.save(user);
-        return new ResponseEntity<>("Added", HttpStatus.OK);
+    public ResponseEntity<String> addUser(User user) throws CarNotFoundException {
+        if (nonNull(user)) {
+            userRepository.save(user);
+            return new ResponseEntity<>("Added", HttpStatus.OK);
+        }
+        throw new CarNotFoundException(String.format("User with %d not found", user.getId()));
     }
+
 }
